Update distributed state
